Test cases for enabling/disabling inline editing.
Calls CKEditor's isInline() to check if the editor has properly enabled
or disabled inline editing mode.

A simple $compile wasn't sufficient for testing the non-inline editing
mode. This mode requires the insertion of elements into the DOM for the
toolbars, which means it had to be added to $document to work in the
test case. This is also why all the other test divs now have the
contenteditable="true" attribute, so that they'll still work with the
original code.